create make promission role name

+++ app/Http/Controllers/UserManagementController.php
@@ -36,45 +36,50 @@ class UserManagementController extends Controller
     {
         DB::beginTransaction();
         try {
-            $user_id       = $request->user_id;
-            $name         = $request->name;
-            $email        = $request->email;
-            $role_name    = $request->role_name;
-            $position     = $request->position;
-            $phone        = $request->phone_number;
-            $department   = $request->department;
-            $status       = $request->status;
+            if (Session::get('role_name') === 'Admin' || Session::get('role_name') === 'Super Admin')
+            {
+                $user_id       = $request->user_id;
+                $name         = $request->name;
+                $email        = $request->email;
+                $role_name    = $request->role_name;
+                $position     = $request->position;
+                $phone        = $request->phone_number;
+                $department   = $request->department;
+                $status       = $request->status;
 
-            $image_name = $request->hidden_avatar;
-            $image = $request->file('avatar');
+                $image_name = $request->hidden_avatar;
+                $image = $request->file('avatar');
 
-            if($image_name =='photo_defaults.jpg') {
-                if ($image != '') {
-                    $image_name = rand() . '.' . $image->getClientOriginalExtension();
-                    $image->move(public_path('/images/'), $image_name);
+                if($image_name =='photo_defaults.jpg') {
+                    if ($image != '') {
+                        $image_name = rand() . '.' . $image->getClientOriginalExtension();
+                        $image->move(public_path('/images/'), $image_name);
+                    }
+                } else {
+                    
+                    if($image != '') {
+                        unlink('images/'.$image_name);
+                        $image_name = rand() . '.' . $image->getClientOriginalExtension();
+                        $image->move(public_path('/images/'), $image_name);
+                    }
                 }
-            } else {
-                
-                if($image != '') {
-                    unlink('images/'.$image_name);
-                    $image_name = rand() . '.' . $image->getClientOriginalExtension();
-                    $image->move(public_path('/images/'), $image_name);
-                }
-            }
             
-            $update = [
-                'user_id'      => $user_id,
-                'name'         => $name,
-                'role_name'    => $role_name,
-                'email'        => $email,
-                'position'     => $position,
-                'phone_number' => $phone,
-                'department'   => $department,
-                'status'       => $status,
-                'avatar'       => $image_name,
-            ];
+                $update = [
+                    'user_id'      => $user_id,
+                    'name'         => $name,
+                    'role_name'    => $role_name,
+                    'email'        => $email,
+                    'position'     => $position,
+                    'phone_number' => $phone,
+                    'department'   => $department,
+                    'status'       => $status,
+                    'avatar'       => $image_name,
+                ];
 
-            User::where('user_id',$request->user_id)->update($update);
+                User::where('user_id',$request->user_id)->update($update);
+            } else {
+                Toastr::error('User update fail :)','Error');
+            }
             DB::commit();
             Toastr::success('User updated successfully :)','Success');
             return redirect()->back();
@@ -91,18 +96,23 @@ class UserManagementController extends Controller
     {
         DB::beginTransaction();
         try {
-
-            if ($request->avatar =='photo_defaults.jpg')
+            if (Session::get('role_name') === 'Super Admin')
             {
-                User::destroy($request->user_id);
+                if ($request->avatar =='photo_defaults.jpg')
+                {
+                    User::destroy($request->user_id);
+                } else {
+                    User::destroy($request->user_id);
+                    unlink('images/'.$request->avatar);
+                }
             } else {
-                User::destroy($request->user_id);
-                unlink('images/'.$request->avatar);
+                Toastr::error('User deleted fail :)','Error');
             }
+
             DB::commit();
             Toastr::success('User deleted successfully :)','Success');
             return redirect()->back();
-            
+    
         } catch(\Exception $e) {
             DB::rollback();
             Toastr::error('User deleted fail :)','Error');
